Update and rename MyWorld.java to TankWorld.java
